+++ load-file-test/nrepl/load_file_sample2.clj
@@ -1,0 +1,3 @@
+(ns nrepl.load-file-sample2)
+
+(Thread/sleep 10000)