+++ src/clojure/nrepl/middleware/interruptible_eval.clj
@@ -121,8 +121,15 @@
                                                      :ns (-> *ns* ns-name str)})))
            ;; TODO: customizable exception prints
            :caught (fn [e]
-                     (let [root-ex (#'clojure.main/root-cause e)]
-                       (when-not (instance? ThreadDeath root-ex)
+                     (let [root-ex (#'clojure.main/root-cause e)
+                           previous-cause (.getCause e)]
+                       ;; Check if the root cause or previous cause of the exception
+                       ;; is a ThreadDeath exception. In case the exception is a
+                       ;; CompilerException, the root cause is not returned by
+                       ;; the root-cause function, so we check the previous cause
+                       ;; instead.
+                       (when-not (or (instance? ThreadDeath root-ex)
+                                     (instance? ThreadDeath previous-cause))
                          (reset! bindings (assoc (capture-thread-bindings) #'*e e))
                          (reset! session (maybe-restore-original-ns @bindings))
                          (t/send transport (response-for msg {:status :eval-error